replaced cos with std::cosf, replaced sin with std::sinf replaced sqrt with std::sqrtf

+++ src/clgame/clg_effects.cpp
@@ -1175,13 +1175,13 @@ void CLG_BigTeleportParticles(vec3_t org)
 
         angle = M_PI * 2 * (rand() & 1023) / 1023.0;
         dist = rand() & 31;
-        p->org[0] = org[0] + cos(angle) * dist;
-        p->vel[0] = cos(angle) * (70 + (rand() & 63));
-        p->accel[0] = -cos(angle) * 100;
+        p->org[0] = org[0] + std::cosf(angle) * dist;
+        p->vel[0] = std::cosf(angle) * (70 + (rand() & 63));
+        p->accel[0] = -std::cosf(angle) * 100;
 
-        p->org[1] = org[1] + sin(angle) * dist;
-        p->vel[1] = sin(angle) * (70 + (rand() & 63));
-        p->accel[1] = -sin(angle) * 100;
+        p->org[1] = org[1] + std::sinf(angle) * dist;
+        p->vel[1] = std::sinf(angle) * (70 + (rand() & 63));
+        p->accel[1] = -std::sinf(angle) * 100;
 
         p->org[2] = org[2] + 8 + (rand() % 90);
         p->vel[2] = -100 + (rand() & 31);
@@ -1565,8 +1565,8 @@ void CLG_OldRailTrail(void)
         VectorClear(p->accel);
 
         d = i * 0.1;
-        c = cos(d);
-        s = sin(d);
+        c = std::cosf(d);
+        s = std::sinf(d);
 
         VectorScale(right, c, dir);
         VectorMA(dir, s, up, dir);
@@ -1687,11 +1687,11 @@ static void CLG_FlyParticles(vec3_t origin, int count)
     ltime = (float)cl->time / 1000.0;
     for (i = 0; i < count; i += 2) {
         angle = ltime * avelocities[i][0];
-        sy = sin(angle);
-        cy = cos(angle);
+        sy = std::sinf(angle);
+        cy = std::cosf(angle);
         angle = ltime * avelocities[i][1];
-        sp = sin(angle);
-        cp = cos(angle);
+        sp = std::sinf(angle);
+        cp = std::cosf(angle);
 
         forward[0] = cp * cy;
         forward[1] = cp * sy;
@@ -1703,7 +1703,7 @@ static void CLG_FlyParticles(vec3_t origin, int count)
 
         p->time = cl->time;
 
-        dist = sin(ltime + i) * 64;
+        dist = std::sinf(ltime + i) * 64;
         p->org[0] = origin[0] + bytedirs[i][0] * dist + forward[0] * BEAMLENGTH;
         p->org[1] = origin[1] + bytedirs[i][1] * dist + forward[1] * BEAMLENGTH;
         p->org[2] = origin[2] + bytedirs[i][2] * dist + forward[2] * BEAMLENGTH;
@@ -1768,11 +1768,11 @@ void CLG_BfgParticles(entity_t* ent)
     ltime = (float)cl->time / 1000.0;
     for (i = 0; i < count; i++) {
         angle = ltime * avelocities[i][0];
-        sy = sin(angle);
-        cy = cos(angle);
+        sy = std::sinf(angle);
+        cy = std::cosf(angle);
         angle = ltime * avelocities[i][1];
-        sp = sin(angle);
-        cp = cos(angle);
+        sp = std::sinf(angle);
+        cp = std::cosf(angle);
 
         forward[0] = cp * cy;
         forward[1] = cp * sy;
@@ -1784,7 +1784,7 @@ void CLG_BfgParticles(entity_t* ent)
 
         p->time = cl->time;
 
-        dist = sin(ltime + i) * 64;
+        dist = std::sinf(ltime + i) * 64;
         p->org[0] = ent->origin[0] + bytedirs[i][0] * dist + forward[0] * BEAMLENGTH;
         p->org[1] = ent->origin[1] + bytedirs[i][1] * dist + forward[1] * BEAMLENGTH;
         p->org[2] = ent->origin[2] + bytedirs[i][2] * dist + forward[2] * BEAMLENGTH;
