+++ src/svgame/physics.cpp
@@ -846,7 +846,7 @@ void SV_Physics_Step(edict_t *ent)
         if ((wasonground) || (ent->flags & (FL_SWIM | FL_FLY)))
             if (!(ent->health <= 0.0 && !M_CheckBottom(ent))) {
                 vel = ent->velocity;
-                speed = sqrt(vel[0] * vel[0] + vel[1] * vel[1]);
+                speed = std::sqrtf(vel[0] * vel[0] + vel[1] * vel[1]);
                 if (speed) {
                     friction = sv_friction;
