+++ src/svgame/weapons/hyperblaster.cpp
@@ -47,9 +47,9 @@ void Weapon_HyperBlaster_Fire(edict_t* ent)
         }
         else {
             rotation = (ent->client->ps.gunframe - 5) * 2 * M_PI / 6;
-            offset[0] = -4 * sin(rotation);
+            offset[0] = -4 * std::sinf(rotation);
             offset[1] = 0;
-            offset[2] = 4 * cos(rotation);
+            offset[2] = 4 * std::cosf(rotation);
 
             if ((ent->client->ps.gunframe == 6) || (ent->client->ps.gunframe == 9))
                 effect = EF_HYPERBLASTER;
